move message codes to light package
