+++ rskj-core/src/main/java/co/rsk/net/light/LightClientMessageCodes.java
@@ -1,4 +1,4 @@
-package co.rsk.net.lc;
+package co.rsk.net.light;
 
 import java.util.HashMap;
 import java.util.Map;
