+++ rskj-core/src/main/java/org/ethereum/net/rlpx/MessageCodesResolver.java
@@ -22,7 +22,7 @@ package org.ethereum.net.rlpx;
 import org.ethereum.net.client.Capability;
 import org.ethereum.net.eth.EthVersion;
 import org.ethereum.net.eth.message.EthMessageCodes;
-import co.rsk.net.lc.LightClientMessageCodes;
+import co.rsk.net.light.LightClientMessageCodes;
 import org.ethereum.net.p2p.P2pMessageCodes;
 
 import java.util.Collections;
